pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "A zero-trust AI gateway."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
httpx = "^0.25.0"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
respx = "^0.20.2"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"